requirements.txt
feedparser
lxml
requests
newspaper3k
nltk
